requirements.txt

speedtest-cli
colorama
